refactor: distance int 타입 변경

+++ src/main/java/nextstep/line/Line.java
@@ -26,7 +26,7 @@ public class Line {
     public Line() {
     }
 
-    public Line(String name, String color, Station upStation, Station downStation, Long distance) {
+    public Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
         this.color = color;
